+++ lib/es5/index.d.ts
@@ -1,3 +1,4 @@
+import { ISasDecodedResult } from './results/sas-decoded-result';
 export declare class Decode {
     /**
      * Convert querystring to name/values pairs. QueryString shouldn't have `?` in it
@@ -8,6 +9,6 @@ export declare class Decode {
      * Decodes Azure Storage SAS token
      *
      */
-    static azureStorageSasToken(sasToken: string): any;
+    static azureStorageSasToken(sasToken: string): ISasDecodedResult;
 }
 //# sourceMappingURL=index.d.ts.map