+++ lib/es5/sasTypes/service-sas-decode-test.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=service-sas-decode-test.d.ts.map