+++ lib/es6/azure-storage-sas-decoded.d.ts
@@ -1,21 +1,4 @@
-/**
- * Type of Sas token
- */
-export declare enum SasType {
-    Account = "Account",
-    Service = "Service",
-    UserDelegation = "UserDelegation",
-    Unknown = "Unknown",
-    Unexpected = "Unexpected"
-}
-export interface SasDecoded {
-    sasToken: string | undefined;
-    sasTokenOptions: Record<string, unknown>;
-    sasType: SasType;
-    results: Record<string, unknown>;
-    formedCorrectly: boolean;
-    error?: Error | undefined;
-}
+import { SasTypeEnum } from './results/sas-decoded-result';
 export default class AzureStorageSasDecoded {
     static decodeSasToken: (sasToken: string, sasTokenOptions: any) => any;
     /**
@@ -27,5 +10,5 @@ export default class AzureStorageSasDecoded {
      *
      * @returns
      */
-    static getSasType(sas: Record<string, unknown>): SasType;
+    static getSasType(sas: Record<string, unknown>): SasTypeEnum;
 }